move cleanup script to it's own file
