+++ packages/loot-core/src/server/budget/goal-template.pegjs
@@ -20,10 +20,6 @@ expr
     { return { type: 'simple', limit , priority: +priority } }
   / priority: priority? _? schedule _ name: name
     { return { type: 'schedule', name, priority: +priority } }
-  / _? source
-    { return { type: 'source' } }
-  / _? sink _? weight: weight?
-    { return { type: 'sink', weight: +weight || 1 } }
 
 repeat 'repeat interval'
   = 'month'i { return { annual: false } }
@@ -50,8 +46,6 @@ starting = 'starting'i
 upTo = 'up'i _ 'to'i
 schedule = 'schedule'i
 priority = '-'i number: number _ {return number}
-source = 'source'
-sink = 'sink'
 
 _ 'space' = ' '+
 d 'digit' = [0-9]
@@ -63,6 +57,5 @@ month 'month' = $(year '-' d d)
 day 'day' = $(d d)
 date = $(month '-' day)
 currencySymbol 'currency symbol' = symbol: . & { return /\p{Sc}/u.test(symbol) }
-weight 'weight' = weight: $(d+) { return +weight }
 
 name 'Name' = $([^\r\n\t]+)
