+++ packages/loot-core/src/server/budget/goaltemplates.ts
@@ -26,134 +26,6 @@ export function overwriteTemplate({ month }) {
   return processTemplate(month, true);
 }
 
-export function cleanupTemplate({ month }) {
-  return processCleanup(month);
-}
-
-async function processCleanup(month) {
-  let num_sources = 0;
-  let num_sinks = 0;
-  let total_weight = 0;
-  let errors = [];
-  let sinkCategory = [];
-  const TEMPLATE_PREFIX = '#cleanup';
-  let category_templates = await getCategoryTemplates(TEMPLATE_PREFIX);
-  let categories = await db.all(
-    'SELECT * FROM v_categories WHERE tombstone = 0',
-  );
-  let db_month = parseInt(month.replace('-', ''));
-  let carryovers = await db.all(
-    `SELECT * FROM zero_budgets WHERE month = ${db_month}`,
-  );
-  let sheetName = monthUtils.sheetForMonth(month);
-  for (let c = 0; c < categories.length; c++) {
-    let category = categories[c];
-    let template = category_templates[category.id];
-    if (template) {
-      if (template.filter(t => t.type === 'source').length > 0) {
-        let balance = await getSheetValue(sheetName, `leftover-${category.id}`);
-        let budgeted = await getSheetValue(sheetName, `budget-${category.id}`);
-        await setBudget({
-          category: category.id,
-          month,
-          amount: budgeted - balance,
-        });
-        num_sources += 1;
-      }
-      if (template.filter(t => t.type === 'sink').length > 0) {
-        sinkCategory.push({ cat: category, temp: template });
-        num_sinks += 1;
-        total_weight += template[0].weight;
-      }
-    }
-  }
-
-  //funds all underfunded categories first unless the overspending rollover is checked
-  for (let c = 0; c < categories.length; c++) {
-    let category = categories[c];
-    let budgetAvailable = await getSheetValue(sheetName, `to-budget`);
-    let balance = await getSheetValue(sheetName, `leftover-${category.id}`);
-    let budgeted = await getSheetValue(sheetName, `budget-${category.id}`);
-    let to_budget = budgeted + Math.abs(balance);
-    let carryover_list = carryovers.filter(c => c.category === category.id);
-    let carryover = 0;
-    if (carryover_list.length > 0) {
-      carryover = carryover_list[0].carryover;
-    }
-    if (
-      balance < 0 &&
-      Math.abs(balance) <= budgetAvailable &&
-      !category.is_income &&
-      carryover === 0
-    ) {
-      await setBudget({
-        category: category.id,
-        month,
-        amount: to_budget,
-      });
-    }
-  }
-
-  let budgetAvailable = await getSheetValue(sheetName, `to-budget`);
-
-  if (budgetAvailable <= 0) {
-    errors.push('No funds are available to reallocate.');
-  }
-
-  for (let c = 0; c < sinkCategory.length; c++) {
-    let budgeted = await getSheetValue(
-      sheetName,
-      `budget-${sinkCategory[c].cat.id}`,
-    );
-    let categoryId = sinkCategory[c].cat.id;
-    let to_budget =
-      budgeted +
-      Math.round(
-        (sinkCategory[c].temp[0].weight / total_weight) * budgetAvailable,
-      );
-    if (c === sinkCategory.length - 1) {
-      let currentBudgetAvailable = await getSheetValue(sheetName, `to-budget`);
-      if (to_budget > currentBudgetAvailable) {
-        to_budget = budgeted + currentBudgetAvailable;
-      }
-    }
-    await setBudget({
-      category: categoryId,
-      month,
-      amount: to_budget,
-    });
-  }
-
-  if (num_sources === 0) {
-    if (errors.length) {
-      return {
-        type: 'error',
-        sticky: true,
-        message: `There were errors interpreting some templates:`,
-        pre: errors.join('\n\n'),
-      };
-    } else {
-      return { type: 'message', message: 'All categories were up to date.' };
-    }
-  } else {
-    let applied = `Successfully returned funds from ${num_sources} ${
-      num_sources === 1 ? 'source' : 'sources'
-    } and funded ${num_sinks} sinking ${num_sinks === 1 ? 'fund' : 'funds'}.`;
-    if (errors.length) {
-      return {
-        sticky: true,
-        message: `${applied} There were errors interpreting some templates:`,
-        pre: errors.join('\n\n'),
-      };
-    } else {
-      return {
-        type: 'message',
-        message: applied,
-      };
-    }
-  }
-}
-
 function checkScheduleTemplates(template) {
   let lowPriority = template[0].priority;
   let errorNotice = false;
@@ -169,8 +41,7 @@ function checkScheduleTemplates(template) {
 async function processTemplate(month, force) {
   let num_applied = 0;
   let errors = [];
-  const TEMPLATE_PREFIX = '#template';
-  let category_templates = await getCategoryTemplates(TEMPLATE_PREFIX);
+  let category_templates = await getCategoryTemplates();
   let lowestPriority = 0;
   let originalCategoryBalance = [];
 
@@ -324,8 +195,8 @@ async function processTemplate(month, force) {
   }
 }
 
-//const TEMPLATE_PREFIX = '#template';
-async function getCategoryTemplates(TEMPLATE_PREFIX) {
+const TEMPLATE_PREFIX = '#template';
+async function getCategoryTemplates() {
   let templates = {};
 
   let notes = await db.all(
